W120 salvage: fix fclCombineEffect.c COMPILE_ERROR (FUN_004bdde0_fcl param 2 void* -> s32); restores MATCH 0 -> 35

+++ src/Event/Fcl/fclCombineList.c
@@ -49,7 +49,7 @@ u64 FUN_003e0870(void);
 u64 FUN_003e08e0(void);
 void FUN_003e0920(void);
 u32 FUN_003e0940(void);
-void FUN_003e0950(u64);
+void FUN_003e0950(s32);
 void FUN_003e0a10(s32);
 void FUN_003e0b70(void);
 u8 FUN_003e0bb0(void);
